shuffled around directory
Added a global folder for our evitics.com webpage.  The site webpages such-as evitics.gatech.edu is hosted in the localsite folder.  When the user logs in he will be directed to the most relevant localsite webapp.  Think of global as the our FAQ, contact us, landing page type of information...it needs to be much-more SEO friendly and faster than the local site ones.