Changed organization list page, and moved user model to include organization collection in heirarchy
The organization list page now shows "My Organizations" and "Other Organizations".  The parsing to check if the user belogns to an org no longer happens in the parse stage. Instead the collection has internal methods which allows for a quick .get(id) check.  I've also added a request pending status for a particular org.